requirements.txt
asgiref==3.8.1
Django==4.2.16
django-bootstrap5==24.3
django-countries==7.6.1
django-environ==0.11.2
pillow==10.4.0
sqlparse==0.5.1
typing-extensions==4.12.2
